1.19.9 gps bedrock clear feature.  package move.
