+++ config_example_1.12/cyclicmagic.cfg
@@ -1254,10 +1254,10 @@ modpacks {
     # Maximum height of the controlled miner block that you can increase it to in the GUI [range: 3 ~ 128, default: 32]
     I:ControlledMiner.maxHeight=32
 
-    # How long the negative potion effects last after a corrupted chorus teleports you [range: 1 ~ 60, default: 10]
+    # How long the negative potion effects last after a corrupted chorus teleports you [range: 0 ~ 600, default: 10]
     I:CorruptedChorusPotions=10
 
-    # How long you can noclip after eating corrupted chorus [range: 1 ~ 60, default: 10]
+    # How long you can noclip after eating corrupted chorus [range: 1 ~ 600, default: 10]
     I:CorruptedChorusSeconds=10
 
     # Max blocks broken by this enchantment at level 1 [range: 1 ~ 512, default: 26]
