+++ src/main/resources/assets/cyclicmagic/lang/en_GB.lang
@@ -1398,7 +1398,7 @@ pump.secondline=<- / ->
 
 item.card_location.name=GPS Marker
 item.card_location.tooltip=Saves the location of a block
-item.card_location.guide=Use on a block to save a location and use in a wireless machine
+item.card_location.guide=Use on a block to save a location and use in a wireless machine.  Use on bedrock to clear item data. 
 item.location.saved=Saved 
 
 wireless.target=Detect target block
