+++ update.json
@@ -324,6 +324,6 @@
 	,"1.19.6":"Fixed bug where flowing water was breaking cables.  "
 	,"1.19.7":"Fixed far left slots of Fluid Transfer Node.  Melter and Solidifier recipes now correctly link in with JEI when looking for uses and recipes for fluids.  zh_CN.lang updated ty sandtechnology.  Dehydrator and fluid tank slightly updated.  Fixed fluid bugs in Solidifier recipes.  Fixed a desync bug with Item Collector's redstone button.  "
 	,"1.19.8":"Glowing Chorus fruit flying no longer counts down while falling or while standing on non-solid blocks.  Glowing Chorus fruit flying effect is now cleansed and removed from player when milk or a Corrupted chorus fruit is eaten.  Trash Void improved better handling of multiple connections at once, should avoid item ducts and other means from backing up.   "
-	,"1.19.9":"Fluid Extraction cables can now pick up water from the world.  Auto User input slots restricted to stack size 1 (fixes several issues, you can still pipe items in as fast as needed).  Exchange scepter now works with silk touch, and can be applied in anvil.  Hydrator & Dehydrator water production compatible with JEI.  The config 'CorruptedChorusPotions' now treats zero as a valid number to disable effect, and higher max value.  "
+	,"1.19.9":"Fluid Extraction cables can now pick up water from the world.  Auto User input slots restricted to stack size 1 (fixes several issues, you can still pipe items in as fast as needed).  Exchange scepter now works with silk touch, and can be applied in anvil.  Hydrator & Dehydrator water production compatible with JEI.  The config 'CorruptedChorusPotions' now treats zero as a valid number to disable effect, and higher max value.  You can delete data in GPS Markers by hitting bedrock.  "
   }
 }
